+++ src/boilrpy/file_generator.py
@@ -1,9 +1,57 @@
+from abc import ABC, abstractmethod
+from typing import Any, Dict, Type
 from boilrpy.file_generators.changelog_generator import ChangelogGenerator
 from boilrpy.file_generators.dockerfile_generator import DockerfileGenerator
 from boilrpy.file_generators.gitignore_generator import GitignoreGenerator
 from boilrpy.file_generators.license_generator import LicenseGenerator
 from boilrpy.file_generators.main_file_generator import MainFileGenerator
 from boilrpy.file_generators.readme_generator import ReadmeGenerator
+from boilrpy.file_generators.pylint_generator import PylintGenerator
+from boilrpy.config import Config
+
+
+class Generator(ABC):
+    """
+    Abstract class for generating project files.
+    """
+    @abstractmethod
+    def generate(self, *args, **kwargs) -> str:
+        """ Generate the content of the file. """
+
+
+class GeneratorFactory:
+    """
+    Factory class for creating generators.
+    """
+    _generators: Dict[str, Type[Generator]] = {
+        "readme": ReadmeGenerator,
+        "license": LicenseGenerator,
+        "gitignore": GitignoreGenerator,
+        "changelog": ChangelogGenerator,
+        "main_file": MainFileGenerator,
+        "dockerfile": DockerfileGenerator,
+        "pylint": PylintGenerator,
+    }
+
+    @classmethod
+    def create_generator(cls, generator_type: str, config: Dict[str, Any]) -> Generator:
+        """
+        Create a generator instance.
+
+        Args:
+            generator_type (str): The type of generator to create.
+            config (Dict[str, Any]): The configuration dictionary.
+
+        Returns:
+            Generator: The created generator instance.
+
+        Raises:
+            ValueError: If the generator type is unknown.
+        """
+        generator_class = cls._generators.get(generator_type)
+        if generator_class is None:
+            raise ValueError(f"Unknown generator type: {generator_type}")
+        return generator_class(config)
 
 
 class FileGenerator:
@@ -11,14 +59,15 @@ class FileGenerator:
     Main class for generating project files.
     """
 
-    def __init__(self, config):
+    def __init__(self, config: Config):
         self.config = config
-        self.readme_generator = ReadmeGenerator(config)
-        self.license_generator = LicenseGenerator(config)
-        self.gitignore_generator = GitignoreGenerator(config)
-        self.changelog_generator = ChangelogGenerator(config)
-        self.main_file_generator = MainFileGenerator(config)
-        self.dockerfile_generator = DockerfileGenerator(config)
+        self.generators: Dict[str, Generator] = {}
+
+    def _get_generator(self, generator_type: str) -> Generator:
+        if generator_type not in self.generators:
+            self.generators[generator_type] = \
+                GeneratorFactory.create_generator(generator_type, self.config)
+        return self.generators[generator_type]
 
     def generate_readme(self, project_info: dict) -> str:
         """
@@ -27,7 +76,7 @@ class FileGenerator:
         :param project_info: Dictionary containing project information
         :return: Content of README.md file
         """
-        return self.readme_generator.generate(project_info)
+        return self._get_generator("readme").generate(project_info)
 
     def generate_license(self, license_name: str, author: str) -> str:
         """
@@ -37,7 +86,7 @@ class FileGenerator:
         :param author: Name of the author
         :return: Content of the license file
         """
-        return self.license_generator.generate(license_name, author)
+        return self._get_generator("license").generate(license_name, author)
 
     def generate_gitignore(self) -> str:
         """
@@ -45,7 +94,7 @@ class FileGenerator:
 
         :return: Content of .gitignore file
         """
-        return self.gitignore_generator.generate()
+        return self._get_generator("gitignore").generate()
 
     def generate_changelog(self, version: str) -> str:
         """
@@ -54,7 +103,7 @@ class FileGenerator:
         :param version: Version number
         :return: Content of CHANGELOG.md file
         """
-        return self.changelog_generator.generate(version)
+        return self._get_generator("changelog").generate(version)
 
     def generate_main_file(self) -> str:
         """
@@ -62,7 +111,7 @@ class FileGenerator:
 
         :return: Content of main.py file
         """
-        return self.main_file_generator.generate()
+        return self._get_generator("main_file").generate()
 
     def generate_dockerfile(self, project_name: str) -> str:
         """
@@ -70,7 +119,7 @@ class FileGenerator:
 
         :return: Content of Dockerfile
         """
-        return self.dockerfile_generator.generate_dockerfile(project_name)
+        return self._get_generator("dockerfile").generate_dockerfile(project_name)
 
     def generate_dockerignore(self) -> str:
         """
@@ -78,4 +127,12 @@ class FileGenerator:
 
         :return: Content of .dockerignore
         """
-        return self.dockerfile_generator.generate_dockerignore()
+        return self._get_generator("dockerfile").generate_dockerignore()
+
+    def generate_pylint(self) -> str:
+        """
+        Generate .pylintrc content.
+
+        :return: Content of .pylintrc
+        """
+        return self._get_generator("pylint").generate()