+++ src/boilrpy/file_generators/readme_generator.py
@@ -30,6 +30,8 @@ ${usage_instructions}
 
 ${testing_instructions}
 
+${pylint_instructions}
+
 ## Contributing
 
 1. Fork the repository
@@ -43,25 +45,30 @@ ${license_info}
 """
         setup_instructions = (
             "### Using Poetry\n\npoetry install"
-            if project_info['use_poetry']
+            if project_info["use_poetry"]
             else "### Using pip and venv\n\npip install -r requirements.txt"
         )
         usage_instructions = (
             "poetry run python main.py"
-            if project_info['use_poetry']
+            if project_info["use_poetry"]
             else "python main.py"
         )
         testing_instructions = (
             "\n## Testing\n\n```poetry run pytest```"
-            if project_info['create_tests'] and project_info['use_poetry']
+            if project_info["create_tests"] and project_info["use_poetry"]
             else "\n## Testing\n\n```pytest```"
-            if project_info['create_tests']
+            if project_info["create_tests"]
+            else ""
+        )
+        pylint_instructions = (
+            "\n## Pylint\n\n```pylint .```"
+            if project_info["use_pylint"]
             else ""
         )
         license_info = (
             f"\n## License\n\nThis project is licensed under the {
-                project_info['license']} License."
-            if project_info['license'] != "None"
+                project_info["license"]} License."
+            if project_info["license"] != "None"
             else ""
         )
         return self.render_template(
@@ -71,5 +78,6 @@ ${license_info}
             setup_instructions=setup_instructions,
             usage_instructions=usage_instructions,
             testing_instructions=testing_instructions,
-            license_info=license_info
+            pylint_instructions=pylint_instructions,
+            license_info=license_info,
         ).lstrip()