+++ tests/file_generators/test_pylint_generator.py
@@ -1,0 +1,11 @@
+import pytest
+from boilrpy.file_generators.pylint_generator import PylintGenerator
+
+
+@pytest.fixture
+def pylint_generator():
+    return PylintGenerator(None)
+
+def test_generate(pylint_generator):
+    result = pylint_generator.generate()
+    assert result == "[MASTER]\ndisable=missing-module-docstring, too-few-public-methods \nignore=tests"